1. common/etcd/go.mod - 修复了空行格式 2. shop-service/config/config.go:30 - 将 TTL 类型从 int 改为 int64，与 etcd.NewServiceRegister 函数的参数类型匹配

+++ shop-service/config/config.go
@@ -27,7 +27,7 @@ type EtcdConfig struct {
 type EtcdServiceConfig struct {
 	Host string `mapstructure:"host"`
 	Name string `mapstructure:"name"`
-	TTL  int    `mapstructure:"ttl"`
+	TTL  int64  `mapstructure:"ttl"`
 }
 
 // ServerConfig 服务器配置
